Jornada Laboral

+++ Software2G/src/com/software2g/contable/facade/impl/GestionFacadeContableSpringService.java
@@ -1207,7 +1207,7 @@ public class GestionFacadeContableSpringService implements IGestionFacadeContabl
 	}
 	
 	@Transactional(propagation=Propagation.NEVER, readOnly=true)
-	public Persona findPersonaById(java.lang.Integer id) throws Exception {
+	public Persona findPersonaById(java.lang.Long id) throws Exception {
 		try {
 			return getPersonaDao().findPersonaById(id);
 		} catch (RuntimeException e) {
